Simplify OpenGL versioned classes
