+++ lib/ome/qtwidgets/CMakeLists.txt
@@ -69,25 +69,15 @@ set(QTWIDGETS_GL_HEADERS
     gl/Image2D.h
     gl/Util.h)
 
-set(QTWIDGETS_GL_V33_SOURCES
-    gl/v33/V33Axis2D.cpp
-    gl/v33/V33Grid2D.cpp
-    gl/v33/V33Image2D.cpp)
+set(QTWIDGETS_GLSL_SOURCES
+    glsl/GLFlatShader2D.cpp
+    glsl/GLImageShader2D.cpp
+    glsl/GLLineShader2D.cpp)
 
-set(QTWIDGETS_GL_V33_HEADERS
-    gl/v33/V33Axis2D.h
-    gl/v33/V33Grid2D.h
-    gl/v33/V33Image2D.h)
-
-set(QTWIDGETS_GLSL_V330_SOURCES
-    glsl/v330/V330GLFlatShader2D.cpp
-    glsl/v330/V330GLImageShader2D.cpp
-    glsl/v330/V330GLLineShader2D.cpp)
-
-set(QTWIDGETS_GLSL_V330_HEADERS
-    glsl/v330/V330GLFlatShader2D.h
-    glsl/v330/V330GLImageShader2D.h
-    glsl/v330/V330GLLineShader2D.h)
+set(QTWIDGETS_GLSL_HEADERS
+    glsl/GLFlatShader2D.h
+    glsl/GLImageShader2D.h
+    glsl/GLLineShader2D.h)
 
 add_library(ome-qtwidgets
             ${QTWIDGETS_SOURCES}
@@ -97,8 +87,8 @@ add_library(ome-qtwidgets
             ${QTWIDGETS_GL_HEADERS}
             ${QTWIDGETS_GL_V33_SOURCES}
             ${QTWIDGETS_GL_V33_HEADERS}
-            ${QTWIDGETS_GLSL_V330_SOURCES}
-            ${QTWIDGETS_GLSL_V330_HEADERS}
+            ${QTWIDGETS_GLSL_SOURCES}
+            ${QTWIDGETS_GLSL_HEADERS}
             ${ome-qtwidgets_HEADERS_MOC}
             ${ome-qtwidgets_RESOURCES})
 
@@ -157,11 +147,8 @@ install(FILES ${OME_QTWIDGETS_HEADERS}
 install(FILES ${OME_QTWIDGETS_GL__HEADERS}
         DESTINATION ${ome_qtwidgets_includedir}/gl
         COMPONENT "development")
-install(FILES ${OME_QTWIDGETS_GL_V33_HEADERS}
-        DESTINATION ${ome_qtwidgets_includedir}/gl/v33
-        COMPONENT "development")
-install(FILES ${OME_QTWIDGETS_GLSL_V330_HEADERS}
-        DESTINATION ${ome_qtwidgets_includedir}/glsl/v330
+install(FILES ${OME_QTWIDGETS_GLSL_HEADERS}
+        DESTINATION ${ome_qtwidgets_includedir}/glsl
         COMPONENT "development")
 
 # Dump header list for testing
